Sharing gone wrong
There was a circular dependecy that TS didn't detect
